stream: Support FileHandle in Read/WriteStream
Include remarks by @addaleax and @aduh95
Move the double inheritance class generator to
the internal event_target
Restore deleted commas
Fix a slightly flaky test

Refs: https://github.com/nodejs/node/pull/35922
