+++ lib/events.js
@@ -28,7 +28,6 @@ const {
   ErrorCaptureStackTrace,
   MathMin,
   NumberIsNaN,
-  Object,
   ObjectCreate,
   ObjectDefineProperty,
   ObjectDefineProperties,
@@ -42,7 +41,7 @@ const {
   String,
   Symbol,
   SymbolFor,
-  SymbolAsyncIterator
+  SymbolAsyncIterator,
 } = primordials;
 const kRejection = SymbolFor('nodejs.rejection');
 
@@ -897,17 +896,3 @@ function on(emitter, event, options) {
     iterator.return();
   }
 }
-
-const EventEmitterMixin = (Superclass) => {
-  class MixedEventEmitter extends Superclass {
-    constructor(...args) {
-      super(...args);
-      EventEmitter.call(this);
-    }
-  }
-  const protoProps = Object.getOwnPropertyDescriptors(EventEmitter.prototype);
-  delete protoProps.constructor;
-  Object.defineProperties(MixedEventEmitter.prototype, protoProps);
-  return MixedEventEmitter;
-};
-module.exports.EventEmitterMixin = EventEmitterMixin;
