+++ lib/internal/fs/promises.js
@@ -20,7 +20,7 @@ const {
   PromisePrototypeThen,
   PromiseResolve,
   Symbol,
-  Uint8Array
+  Uint8Array,
 } = primordials;
 
 const {
@@ -71,7 +71,7 @@ const {
 } = require('internal/validators');
 const pathModule = require('path');
 const { promisify } = require('internal/util');
-const EventEmitter = require('events');
+const { EventEmitterMixin } = require('internal/event_target');
 
 const kHandle = Symbol('kHandle');
 const kFd = Symbol('kFd');
@@ -97,7 +97,7 @@ const lazyDOMException = hideStackFrames((message, name) => {
   return new DOMException(message, name);
 });
 
-class FileHandle extends EventEmitter.EventEmitterMixin(JSTransferable) {
+class FileHandle extends EventEmitterMixin(JSTransferable) {
   constructor(filehandle) {
     super();
     this[kHandle] = filehandle;
@@ -229,7 +229,6 @@ class FileHandle extends EventEmitter.EventEmitterMixin(JSTransferable) {
   [kDeserialize]({ handle }) {
     this[kHandle] = handle;
     this[kFd] = handle.fd;
-    EventEmitter.call(this);
   }
 
   [kRef]() {
@@ -727,5 +726,5 @@ module.exports = {
 
   FileHandle,
   kRef,
-  kUnref
+  kUnref,
 };
