+++ lib/internal/fs/streams.js
@@ -14,7 +14,7 @@ const {
 const {
   ERR_INVALID_ARG_TYPE,
   ERR_OUT_OF_RANGE,
-  ERR_METHOD_NOT_IMPLEMENTED
+  ERR_METHOD_NOT_IMPLEMENTED,
 } = require('internal/errors').codes;
 const { deprecate } = require('internal/util');
 const { validateInteger } = require('internal/validators');
